add locgic and icons for network detector
