+++ README.md
@@ -29,3 +29,9 @@ My personal fun website that is a super overkill MERN stack
 Icons made by <a href="https://www.flaticon.com/authors/freepik" title="Freepik">Freepik</a> from <a href="https://www.flaticon.com/"             title="Flaticon">www.flaticon.com</a> is licensed by <a href="http://creativecommons.org/licenses/by/3.0/"             title="Creative Commons BY 3.0" target="_blank">CC 3.0 BY</a></div>
 
 <a href="https://www.freepik.com/free-photos-vectors/abstract">Abstract vector created by freepik - www.freepik.com</a>
+
+<a target="_blank" href="https://icons8.com/icons/set/wi-fi-connected">Wi-Fi Connected</a> icon by <a target="_blank" href="https://icons8.com">Icons8</a>
+
+<a target="_blank" href="https://icons8.com/icons/set/wifi-off">Wi-Fi off</a> icon by <a target="_blank" href="https://icons8.com">Icons8</a>
+
+<a target="_blank" href="https://icons8.com/icons/set/wi-fi-disconnected">Wi-Fi Disconnected</a> icon by <a target="_blank" href="https://icons8.com">Icons8</a>
