+++ src/App.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent, Fragment } from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import Loadable from 'react-loadable';
 
-import NetworkDetector from './shared/network-detector/network-detector';
+import NetworkStatus from './shared/network-status/network-status';
 
 import styles from './app.module.scss';
 import Routing from './routing';
@@ -16,12 +16,12 @@ const App: FunctionComponent = () => {
   return (
     <Fragment>
       <LazyStarfield />
+      <NetworkStatus />
       <div className={styles.content}>
         <BrowserRouter>
           <Routing />
         </BrowserRouter>
       </div>
-      <NetworkDetector />
     </Fragment>
   );
 };
