+++ src/welcome/welcome.module.scss
@@ -20,24 +20,27 @@
 
 .icon-holder {
   position: relative;
-  width: 50px;
+  width: 30px;
   height: 50px;
-  svg {
+  .rocket {
+    transform: rotate(45deg);
     z-index: 2;
     top: 0;
     left: 0;
     position: absolute;
+    height: 50px;
+    width: auto;
   }
-  svg + svg {
+  .rocket + .rocket {
     opacity: 0;
   }
 }
 
 .flame {
   position: relative;
-  top: 65px;
-  left: -15px;
+  top: 55px;
+  left: -16px;
   z-index: 1;
-  width: 80px;
+  width: 60px;
   height: 0px;
 }
